+++ packages/extensions/auracle-pack/src/index.tsx
@@ -1,5 +1,4 @@
 import { AuracleStatusChip } from './components/StatusChip';
-import { AuracleConnections } from './components/ConnectionsSettings';
 import { AuracleFlowEditor } from './components/FlowEditor';
 import { GridPanel } from './components/grid/GridPanel';
 import { GridGutterButton } from './components/grid/GridGutterButton';
@@ -12,10 +11,6 @@ export async function deactivate() {}
 
 export const hostComponents = {
   AuracleStatusChip,
-};
-
-export const settingsPanel = {
-  AuracleConnections,
 };
 
 // ONE rail surface: the Grid. Every panel the pack used to register is a room